requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1
django-debug-toolbar==3.6.0
django-filter==22.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
Faker==14.2.0
idna==3.3
inflection==0.5.1
jsonschema==4.15.0
packaging==21.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
